fix(reply-composer): issue #89; Refactored reply composer logic into a separate component
